- Add JSONService to decode JSON - Add unit tests for decoding to models (check for no errors) - Fix issues with decoding   - Add optionals or CodingKeys   - Remove CodingKeys for related Enums and conform to String   - Add missing fields to related Enums - Add SubPoisonMist and SubPointSensor

+++ Sources/Splatalyzer/Models/Game Parameters/Brellas/BrellaBulletShotgunGroupParameter.swift
@@ -11,7 +11,7 @@ struct BrellaBulletShotgunGroupParameter: Codable {
     let addDegree: Double?
     let collisionParam: BrellaBulletShotgunGroupCollisionParameter?
     let damageParam: BrellaBulletShotgunGroupDamageParameter
-    let drawRadius: Int?
+    let drawRadius: Double?
     let horizontalDegree: Double
     let horizontalOffset: Double
     let initVelRandom: Double?
